feat: mv crud.py the right place
