+++ src/app.py
@@ -181,113 +181,16 @@ ref: https://fastapi.tiangolo.com/tutorial/security/oauth2-jwt/
 User: john
 Password: secret
 
-alice/secret2
-
 """
 
 # to get a string like this run:
-# openssl rand -hex 32
+# generate CMD: openssl rand -hex 32
+# TODO: 適切なファイルに移動すろ
 SECRET_KEY = "ea7532c3afb46f1129d620ebfae641119511c10dc62fcd3ce33722cce56938a9"
 ALGORITHM = "HS256"
 ACCESS_TOKEN_EXPIRE_MINUTES = 30
 
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class TokenData(BaseModel):
-    username: Optional[str] = None
-
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
-
-
-def verify_password(plain_password, hashed_password):
-    """
-
-    :param plain_password:
-    :param hashed_password:
-    :return:
-    """
-    return pwd_context.verify(plain_password, hashed_password)
-
-
-def get_user(db, username: str):
-    """
-    >>> db: dict
-    {
-     'john': {
-       'username': 'john',
-       'full_name': 'John Doe',
-       'email': 'johndoe@example.com',
-       'hashed_password': '$2b$12$EixZaYVK1fsbw1ZfbX3OXePaWxn96p36WQoeG6Lruj3vjPGga31lW',
-       'disabled': False
-      }
-    }
-
-    >>> db[username]: dict
-    {
-     'username': 'john',
-     'full_name': 'John Doe',
-     'email': 'johndoe@example.com',
-     'hashed_password': '$2b$12$EixZaYVK1fsbw1ZfbX3OXePaWxn96p36WQoeG6Lruj3vjPGga31lW',
-     'disabled': False
-    }
-
-    :param db:
-    :param username:
-    :return:
-    """
-    # TODO: mock user
-    user = crud.get_user(db, user_id=2)
-    username_db = {
-        username: {
-            'username': user.username,
-            'email': user.email,
-            'hashed_password': user.hashed_password,
-            'disabled': user.disabled,
-        }
-    }
-    if username in username_db:
-        user_dict = username_db[username]
-        return schemas.UserInDB(**user_dict)
-
-
-def authenticate_user(db, username: str, password: str):
-    """
-    Note:
-      (Pdb) verify_password(password, user.hashed_password)
-      *** passlib.exc.UnknownHashError: hash could not be identified
-    """
-    user = get_user(db, username)
-    if not user:
-        return False
-    pdb.set_trace()
-    if not verify_password(password, user.hashed_password):
-        return False
-    return user
-
-
-def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
-    """
-
-    :param data: {'sub': 'john'}
-    :param expires_delta: datetime.timedelta(seconds=1800)
-    :return:
-    """
-    pdb.set_trace()
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
 
 
 async def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
@@ -302,30 +205,28 @@ async def get_current_user(token: str = Depends(oauth2_scheme), db: Session = De
         username: str = payload.get("sub")
         if username is None:
             raise credentials_exception
-        token_data = TokenData(username=username)
+        token_data = schemas.TokenData(username=username)
     except JWTError:
         raise credentials_exception
-    user = get_user(db, username=token_data.username)
+    user = crud.get_username(db, username=token_data.username)
     if user is None:
         raise credentials_exception
     return user
 
 
 async def get_current_active_user(current_user: schemas.User = Depends(get_current_user)):
-    print('current_user: L309', current_user)
     if current_user.disabled:
         raise HTTPException(status_code=400, detail="Inactive user")
     return current_user
 
 
-@app.post("/token", response_model=Token)
+@app.post("/token", response_model=schemas.Token)
 async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(),
                                  db: Session = Depends(get_db)
                                  ):
     """ API: login
-    TODO: fake_user_db -> db
     """
-    user = authenticate_user(db, form_data.username, form_data.password)
+    user = crud.authenticate_user(db, form_data.username, form_data.password)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
@@ -333,7 +234,7 @@ async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(
             headers={"WWW-Authenticate": "Bearer"},
         )
     access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    access_token = create_access_token(
+    access_token = crud.create_access_token(
         data={"sub": user.username}, expires_delta=access_token_expires
     )
     return {"access_token": access_token, "token_type": "bearer"}
