+++ src/schemas.py
@@ -43,3 +43,12 @@ class User(UserBase):
 
 class UserInDB(User):
     hashed_password: str
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class TokenData(BaseModel):
+    username: Optional[str] = None
